[FIX] account: bank statement currency
If there is no currency set on the journal,
the fallback currency must be the currency
of the journal company, not the currency
of the user company.

This would mean that when you change of company,
with another currency set,
the bank statement change of currency, this is non-sense.

opw-670187