[IMP] base,openerp,web: add debug mode === 'assets' and don't split assets by default
This is a rather large change for developers: from this commit on, we
need to use debug=assets instead of just debug to split assets.

The About menu has been slightly improved as well (the link does not
appear when we are already in debug mode), and an option has been added
in the debugging menu to activate the assets debugging mode.