[FIX] l10n_multilang: create translations in master lang
When creating the account translations using the account template translations
read it without the environment language.

Otherwise, when the translation is created, the content of `element.name` (the
source term to translate) might be in a different language and already
translated.

This was introduce during the migration at the new API where previous code did
not propagate the context (probably by accident).