[IMP] website_sale: clean cart update methods
* _cart_find_product_line

The method _cart_find_product_line is always called with only one record.
The fact that the return statement was inside the loop confirms the for loop is
not useful.

The method will now return recordsets instead of ids.
Convert the overrides to the new API for cleaner code and avoid old-new api
headhache.

Side note: the website_sale_option override logic is very counter intuitive and
should probably be refactored.

* _cart_update

Add ensure_one to be sure working on only one sale.order.

The method is always called on only one order and does not make sense to be used
on multiple (is the action that is done on one line of the cart).
The return format is the updated line without link to multiple orders.

Remove the for loop and assume self is only one order.