[FIX] sale: `Quotations` filter in the sales report
The `quotations` filter in the sales report did not
take into account the `sent` sale orders.

This is following the revision
7f5f945dcfc81e39d36eac06009bee109721c471
which meant sense in Odoo 8.0, as
this report is based on the `sale.order.line`,
and the state on the line could be different
than the state on the order, and there was no
`sent` state anyway among the possible values
of the line state selection.

But, for Odoo 9.0, the state on the line
is a related to the state of the order,
and `sent` must therefore be taken
into account again.

opw-671719