[FIX] account: settings, prevent the re-write of the company currency
When applying the accounting configuration wizard,
the `currency_id` was being re-written on the company,
even if there was no change (if the currency remained the same),
because of the related.

This leaded as well to the re-write of the company currency
on all journal items, because the `company_currency_id` on
`account.move.line` is a related to `company_id.currency_id`as well,
and they were being recomputed due to the above re-write.

Meaning, when you had a bunch of journal items, applying the
accounting configuration wizard could have take a while.

opw-672400