[FIX] sales_team: allow to pass False as default saleteam
Before this fix, sales_team.team_sales_department was always used ignoring default_ in context.
Now we check before in context if a default sale team is forced.
We allow False as saleteam, eg you assign a lead to a employee who is not a saleman.

Now we check that the default saleteam sales_team.team_sales_department is configured to use_leads
to avoid undefined saleteam in the front end.